Restructure into multiple files
And in the process clean up the date handling in TogglWeekly.
